requirements.txt
google-genai>=0.3.0
openai>=1.1.1
openlit
fastapi
uvicorn[standard]
pydantic